% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhois-package.R
\docType{package}
\name{pwhois}
\alias{pwhois}
\alias{pwhois-package}
\title{pwhois}
\description{
WHOIS (pronounced as the phrase "who is") is a query and response protocol
that is widely used for querying databases that store the registered users
or assignees of an Internet resource, such as a domain name, an IP address
block or an autonomous system, but is also used for a wider range of other
information. The protocol stores and delivers database content in a human-
readable format. The current iteration of the WHOIS protocol was drafted
by the Internet Society, and is documented in RFC 3912. Tools are provided
to issue query requests and process query responses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/pwhois}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/pwhois/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
