% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_cymru}
\alias{tidy_cymru}
\title{Turn a Team Cymru server response into a data frame}
\usage{
tidy_cymru(cymru_raw_response)
}
\arguments{
\item{cymru_raw_response}{the raw response from a call to \code{\link[=whois]{whois()}}}
}
\value{
data frame
}
\description{
Turn a Team Cymru server response into a data frame
}
\examples{
res <- whois('type=cymru 17.253.144.10') # 'type=cymru' is only supported on pwhois
tidy_pwhois(res)
}
