% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_pwhois}
\alias{tidy_pwhois}
\title{Turn a Prefix WhoIs server response into a data frame}
\usage{
tidy_pwhois(pwhois_raw_response)
}
\arguments{
\item{pwhois_raw_response}{the raw response from a call to \code{\link[=whois]{whois()}}}
}
\value{
data frame
}
\description{
Turn a Prefix WhoIs server response into a data frame
}
\examples{
res <- whois("registry org-name=apple, inc") # 'registry' is only supported on pwhois
tidy_pwhois(res)
}
