% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-whois.R
\name{whois_query}
\alias{whois_query}
\title{Go WHOIS}
\usage{
whois_query(query, server = "")
}
\arguments{
\item{query}{WHOIS query}

\item{server}{leave \code{""} for the default}
}
\description{
Go WHOIS
}
