% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhois-package.R
\docType{package}
\name{pwhois}
\alias{pwhois}
\alias{pwhois-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/pwhois}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/pwhois/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
